Set 3.10.2.0 as recommended to unbreak windows
Related:

* https://github.com/haskell/ghcup-hs/issues/966

We bump this, although 'cabal run' is known to be broken
on windows: https://github.com/haskell/cabal/issues/9334

However, backporting more patches to 3.6 is not sustainable.
Initially, I tried to backport
https://github.com/haskell/cabal/pull/8071,
but it's not clear whether that is the only thing that causes
compatibility issues with GHC 9.4 and might drive us into
an awkward situation as well.