Add OpenBSD as an explicitly supported platform
GHCup currently doesn't run on OpenBSD (or any platform not explicitly
supported), even if it builds. This adds explicit support for OpenBSD.

We may later not block it from running on unsupported platforms, even
when we don't explicitly make sure our build scripts work for them or
provide bindists or CI for them.